requirements.txt
# Core Dependencies
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.0
openai>=1.0.0

# SSH and Remote Execution
ssh2-python>=1.0.0
paramiko>=3.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Email (for Gmail notifications)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0